feat: Update Dockerfile to use Python 3.12.4-slim image
This commit adds a new Dockerfile that uses the official Python 3.12.4-slim image as the parent image. It sets the working directory to `/usr/src/app` and copies the contents of the current directory into the container. The `requirements.txt` file is used to install the necessary packages, and the `calendarbot.py` file is set as the entry point for the container.